+++ tools/output/null.rs
@@ -533,13 +533,13 @@ pub struct Muxer {
 pub type NullOutputContext = MuxerPriv;
 unsafe extern "C" fn null_write(c: *mut NullOutputContext, p: *mut Dav1dPicture) -> libc::c_int {
     dav1d_picture_unref(p);
-    return 0 as libc::c_int;
+    return 0i32;
 }
 #[no_mangle]
 pub static mut null_muxer: Muxer = unsafe {
     {
         let mut init = Muxer {
-            priv_data_size: 0 as libc::c_int,
+            priv_data_size: 0i32,
             name: b"null\0" as *const u8 as *const libc::c_char,
             extension: b"null\0" as *const u8 as *const libc::c_char,
             write_header: None,
